Add partial shim for libcurl in internet module
Reimplements some of the libcurl functions originally in internet.so
using a shim to call out to the emscripten fetch API.

The default forms of url() and download.file() now work in webR, subject
to HTTP access control (CORS) rules. The shim downloads the files using XHR,
polling on the main thread until they have finished or errored out.

XHR is asynchronous, but we can't fully block on the main thread to wait
for a download to complete, so ASYNCIFY has been enabled as part of the
emscripten build. Since this increases binary size, the optimisations
have been changed to -Oz for more aggressive space trimming.

ASYNCIFY in a side module does not work currently, so the functions
have been static linked into the main module for now.

Enabling ASYNCIFY causes stack errors on Safari, so the JIT compiler
within R has been disabled using an environment variable, which seems to
help. Possibly the threaded code generated by the compiler leads to large
stack use. In any case, more work needs to be done to reduce the amount
of stack used in webR so that it is more stable. The default behaviour can
be re-enabled by loading the compiler library and running enableJIT(3).