Build R using Makefiles and split up R patches
This commit switches the way the build_R script works so that R's own
internal build system is used. This should avoid the need of a new
Makefile for webR.

The stage1 and stage2 patches have been split up into separate files,
managed using quilt.

Some changes to emfc are included:
    * Find some extra vararg symbols to be patched
    * Find and rename the cabs and csqrt fortran intrinsics
Later we will need to reimplement the intrinsics, but this should at
least avoid the conflict between the musl and fortran version of the
functions.

Some issues to work on:
    * Some packages, such as cli, are now broken with this change
    * Plotting is broken. We need to compile devCanvas.c, but we can't
    just link it with grDevices.so because ASM_JS doesn't work in side
    modules yet. Maybe we can pass the string to a special function
    linked into the main module?