Tweak R_jumpctxt value, revert ctxt level changes
This is a smaller fix to the issue in webR when context jumping during
an error.

A value of 0 is not a valid value to pass to longjmp. The man page
states that if longjmp is invoked with a second argument of 0, 1
will be returned instead. This seems to work in Linux, but not with
Emscripten. So, this patch changes the value passed to longjmp when
calling jump_to_top_ex to the value of 1 explicitly.