Tweak R build script
Rebuilding R one final time at the end of the script seems to help with
the addTaskCallback problem.

This could still be improved, at the moment a native version of R needs
to be built and used in the built process, so that packages can be made
lazy loading. This could perhaps be separated out so that a non-native
R can be built and packages made lazy loading more independently.