Fix R plotting in web worker
The canvas R graphics device is modified to pass its operations as
messages in the output queue. The application can listen for these
messages and run the Canvas API calls appropriately.

Probably in future we should refactor this patch entirely to strip out
all the canvas specific calls and just pass the generic R graphics device
calls to the application using webR. The application can then implement
the R graphics device API itself, drawing to a Canvas element or otherwise.