Handle R SEXP objects via an RProxy object
The new R function Module._evalRCode returns SEXP pointers. In this commit
the pointers are wrapped in an RProxy object in the worker thread using
the function convertSEXP. This function uses the SEXP type to decide which
subclass of RProxy to create, with a different subclass for each SEXP type.

Some SEXP types should be able to be converted to JavaScript objects without
loss of structure. A property on the SEXP subclass notes if such an
implicit conversion should be made, and convertSEXP performs the conversion
automatically using the toJs method if so. convertSEXP returns either the
converted object or a RProxy object.

In this commit, the RProxy handlers do not do much, simply passing on get()
requests to the underlying subclass target. In future, the handler(s) should
be extended to allow for accessing R object attributes on complicated SEXP
types as if they were JS properties, calling R functions via the RProxy in
JS, deal with R promises, etc.

In this commit, work has begun for the following SEXP types, which have been
setup for implicit conversion:

 * NILSXP
 * INTSXP
 * REALSXP
 * CPLXSXP
 * STRSXP
 * RAWSXP