Move RProxy methods to main thread
When evaluating R code the resulting RProxy objects are now created
on the main thread. R SEXP objects are either implicitly converted
to JS types before being passed back from the worker, or otherwise
the raw SEXP pointer is passed back.

The main thread RProxy object targets a SEXP pointer, which is passed
back to the worker as part of trapped requests for properties as a
form of ID. The worker then returns properties as another pointer to
a SEXP object or implicitly converted into a transferrable type.