Tweak RProxy infrastructure, add setter handler
Methods on the RSexp classes handle getting and setting via index
for vector types and and optionally the `names` attribute. In this
commit that work has begun with the RSexpReal class, but futher
additions are still required for the other classes of SEXP.

Argument handling is dealt with a little better using a createRSexp
function (perhaps this should be done via the RSep constructor).

Arguments of type scalar number and scalar string can be passed
through to R. R SEXP objects are created, so they should be subject
to the normal R garbage collector.