Add tests based on R's testInstalledPackage output
These scripts are based on the output of `tools::testInstalledPackage`
for each of R's default packages. The test scripts have been modified to
remove obviously broken tests, such as those that depend on help content
or other files that are not packaged as part of the webR virtual
filesystem (e.g. COPYING, TESTING, NEWS, etc.)