Prompt user for downloading webR binary packages
In the demo REPL app, install a handler for packageNotFoundError that
prompts the user if they would like to download a webR binary package.

If the user replies in the negative, the package error is thrown as
it would normally.

If the user replies in the affirmative, the package is downloaded using
webr::install. Once that has finished, the "retry_loadNamespace" restart
is invoked if it exists, continuing execution. Otherwise the "abort"
restart is invoked so as to return to the top level, squelching the
package error message.