Add option to force R to start non-interactively
This will become useful when testing R packages so that we can avoid
asking for interactive input, such as "Hit <Return> to see next plot".

This option should be documented with a warning, starting with
interactive=false means that while evalRCode works as expected using
R via the REPL will no longer be possible.