Add unpackScalarArrays function to utils.ts
This function searches through a JS object recursively looking for
Arrays with only 1 element. An object is returned with such Arrays
replaced with their first element, here referred to as "unpacking"
the Array.

The function is used in evalRCode, including a minor change to
protect the object during conversion to JS. Unit tests are also
updated to reflect the unpacked arrays.