Reimplement JS structure for lists and vectors
 * toArrays() returns list and vector objects in the form,
    { names: [ 'a', 'b', ... ], values: [ x, y, ...] }

 * toArray() returns the values of a list or vector object,
    [ x, y, ... ]

 * toObject() returns an object with keys based on names,
    { 'a': x, 'b': y, ... }

 * entries() returns a "zipped" object in a format compatible with
   Object.fromEntries(),
    [ [ 'a', x ], [ 'b', y ], ...]

For atomic vectors,

 * toTypedArray() returns a typed array, e.g. Float32Array
 * missing() returns an array of indicators for missing values,
    [ false, false, false, true, false, ... ]
 * toArrays() is as above, but includes a 'missing' key, containing the
   result of missing().

When computing names for toArrays() and names(), missing names are
replaced with undefined in the array. For a list or vector with no names
at all the entire names array is replaced with undefined.

When computing names for entries() and toObject(), missing names are
replaced with the index, as used by R. For example, when the R object
    list(a="x", "y")
is converted to a JS object with toObject() it becomes,
    { 2: 'y', a: 'x' }

When converting to JS objects using lists with duplicate names, the first
instance of the name wins, unless the option throwOnDuplicateKey is passed
to toObject(), in which case an error is thrown.