Detect and handle an interrupt signal for R
A userBreakSignal property is created on the main thread as a way to
signal to R to interrupt computation when it is non-zero. The property
can be set with `webR.interrupt()`.

The worker thread checks the value of this property periodically via
a `SyncTask` in `handleEvents()`. If the value is non-zero, R's
`onintr()` is called to interrupt the current computation and return
to the main prompt.

Later, `handleEvents()` can be expanded to handle other types of events
on the main thread.

The webR Console class is updated to include an `interrupt` method. The
webR REPL app is updated to set the userBreakSignal property on Ctrl-C.