RProxify nested RTargetObjs from worker thread
When getting results from the worker thread, newRproxy now finds and
replaces all RTargetObjs in the returned object, rather than just
checking if the object itself is an RTargetObj.

This allows one to return more complicated structures containing proxies
to R objects, such as an Array of RObjects or a JS object with nested
RObjects.

The corresponding RProxy & RProxify type mappings are updated to reflect
the change.