Update toTree, toArray, toObject for depth
For R Lists, Pairlists and Environments, the depth argument can be
provided to the methods. webR will only convert to JS objects up to
the nested depth given.

Objects that are deeper than the given depth parameter are returned as
RObjImpl objects on the worker thread, and so RObjects on the main thread.