Send main thread clientID on service worker req
In Safari, sending a message from the service worker to the client via
postMessage sends a message to the worker thread, rather than the main
thread as intended. In Chrome, however, the message goes to the main
thread by default.

Since the message is intended to be consumed and replied to by the main
thread while the worker blocks, we include the client ID for the main
thread in out service worker XHR fetch request. With this way, we can
explicitly ask the service worker to postMessage to the main thread
while the worker blocks.

This fixes the service worker channel in Safari.