Better handle targets in RObjImpl constructor
The RObjImpl constructor is changed so that it takes either an
RTargetObj or RawType. RTargetObj of type PTR are wrapped, and
objects of type RAW or raw objects passed directly to the constructor
are converted into a new object via newRObjFromTarget().

When constructing instances of RObjImpl for pre-existing R objects,
RObjImpl.wrap() should be used to ensure that the correct subclass is
instantiated. Using `new RObjImpl()` with an RTargetPtr will work, but
only the base class methods will be available for use.