Transfer data using postMessage in service worker
Before performing a sync XHR in the service worker channel, first send
the request data using postMessage. This leaves open the possibility to
transferring objects directly from the worker thread to the main thread
as part of the request.

Encoding and decoding the data that is sent over the XHR (e.g. the
response from the main thread to the worker thread) has been abstracted
into encodeData() and decodeData() functions. These functions are now
used by both the service worker channel and the shared arraybuffer channel
to encode data for the wire. This ensures the data encoding is the same
for both channels and any later introduced replacer functions only need
to be added in a single place.