Proxy RObjectTree when transferring to main thread
The RObjectTree type is reorganised so that on the worker thread tree
object types are in terms of RObjImpl type, while on the main
thread they are in terms of the RObject type.

For atomic vectors where the objects cannot be nested, a
RObjectTreeAtomic type is created that takes the base type as a
parameter.