Don't add missing property during atomic toTree
At the moment the default is to use `toArray()` to populate the
values property when converting atomic vectors to JavaScript. Since
`toArray()` returns `null` for missing values unambiguously, there
is no need to include the missing property as part of the result.

The missing property remains optional on the `RObjectTree` type in
preparation for a possible future option introduced so that `toTree()`
returns a JavaScript typed array for the values property instead,
which would include the raw sentinel values so indeed need a separate
array of missing indicators.