Don't convert evalRCode captured conditions to JS
This avoids an issue where evalRCode captures a condition with a complex
call object in the condition list. Attempting to convert this fully to
JS can fail if it contains objects like closures and functions.

Instead return captured conditions in the evalRCode output list in the
form:

  { type: 'error', data: RObject }

where RObject is an R proxy object associated with the list object for
the raised condition.