Backport R's strict-prototypes fix for dotcall functions (#103)
* Backport strict-prototypes fix for dotcall

R source has committed a fix for the strict-prototypes problem with R's
dotcall functions that was previously worked around for webR with a patch
to trampoline through JavaScript. The R fix has been backported to the
R-4.1.3 patches for webR.

* Fix do_dotCode with compatible Fortran prototypes

Add typedefs for casting Fortran symbols to function prototypes
that are compatible when running under Emscripten. With this, Fortran
symbols are now cast to the form,

void fun(void**, [...], void**)

i.e. without a return type and with arguments passed in by reference.