Reorganise the type used for serialising R objects
The RObjTree type is reorganised and expanded to better support more
types of R object. The different serialised forms for R NULL, string,
symbol and atomic vector types are formalised, and the naming for the
types making up the RObjTree union is tidied up. Expanding the
JavaScript object contents for serialised R objects should be simpler
after this change.

By hard setting the particular strings in the type property, TypeScript
can now better infer what properties are available for R objects serialised
with functions such as toTree(), and better understands how objects are
nested using the RObjTreeNode type.

Typedoc documentation is added explaining the meaning of the RObjTree
and RObjData types, and how R object references are handled on the main
and worker threads by a type parameter.