Use R_tryCatchError when evaluating R calls
This ensures that the stack is only unwound as far as the calling
function, rather than all the way up to the top level, before throwing a
JS Error in the worker thread.

This fixes an issue where JS `finally` blocks would continue to try to
unprotect objects after R has already jumped to top level and reset the
protection counter.