Simplify `safeEval()`
Need to add `getWasmTableEntry()` to `EXPORTED_FUNCTIONS`.
This causes this warning:

```
EXPORTED_FUNCTIONS is not valid with LINKABLE set (normally due to SIDE_MODULE=1/MAIN_MODULE=1) since all functions are exported this mode.
```

But otherwise it doesn't get included in `Module`. Trying to add it to
`EXPORTED_RUNTIME_METHODS` doesn't work and causes this warning:

```
warning: invalid item in EXPORTED_RUNTIME_METHODS: getWasmTableEntry
```