Fix and simplify the webr test_package function
The error handlers in this function contained a bug in that the
`cond` argument was missing. The argument has been added and the
logic simplified so that if an error occurs in the test code it is
caught and re-thrown in the context of the `test_package` function.

With this, problems in test code are not absorbed by the `sink()` and
so can be seen directly in the output when testing webR with
e.g. `make check-pr`.

Also removes a patch that skips a test that seems to work OK now.