Run fc-cache as part of the Wasm fontconfig build
By caching information about fonts in `/usr/share/fonts` at build time, webR
does not need to scan the directory at runtime to generate a list of fonts.
When using a lazy filesystem, this avoids downloading unused fonts
unnecessarily.

Font config's cache is stored at `/var/cache/fontconfig` in the virtual
filesystem.

For this to work well fontconfig must be patched so that,
 * `fc-config` is able to run on the host system through node to generate
   cache files for use in Wasm.
 * The cache files are accepted as valid, even though the timestamp of
   the `/usr/share/fonts` directory in the virtual file system has changed.