Do not minify esbuild output of worker scripts
It is expected by webR package shims such as the httpuv and later
packages that certain JS objects and functions are accessible in
the global scope. For example, such functions are used to communicate
with the main thread through the webR communication channel.

When bundling is enabled, the esbuild minifier assumes that its output
will only be accessed by objects within its own scope and so minifies
the names of even top-level declarations. So, for the moment minifying
has been disabled for the worker script outputs.

This is a return to the behaviour of the `esbuild.js` script in use
before #224.