Tweak npm package build recipe
The recipe copying the content of the virtual filesystem into place
for building the webr npm package has been combined with the `package` target
in `src/Makefile`.

Since the `package` target is declared PHONY, this ensures that the virtual
filesystem content is always made up to date when building an npm package.

The issue has become visible after #224 due to changes in the way the
development server has been implemented and is run.