Don't add the Pragma HTTP header for XHR requests
The Pragma HTTP header is deprecated[1] and when used with Cross-Origin
URLs leads to a non-simple request. Non-simple requests trigger a CORS
preflight check, which breaks when the web server does not support HTTP
`OPTIONS` requests, such as with GitHub Pages or a local server started
using httpuv.

Previously, the header was used when the `cacheOK` argument was set to
`FALSE`. However, in the Wasm browser environment it probably makes
more sense to delegate the decision of caching to the web browser's own
cross-origin request mechanism, so the `cacheOK` argument is now ignored
when making XHR requests.

By default, a cached but working response is probably better than no
response at all, but if such a "no-cache" setting is required and the
server supports OPTIONS, it may still be set manually by the end-user
through the `headers` argument.

[1] https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Pragma
[2] https://developer.mozilla.org/en-US/docs/Web/HTTP/CORS#simple_requests